+++ ChangeLog.md
@@ -11,6 +11,9 @@
   options that allow the Pygments syntax-highlighting colour scheme to be
   configured independently for light and dark mode.
   ([#334](https://github.com/davep/blogmore/pull/334))
+- Fixed `serve` mode not restoring configuration defaults when a setting is
+  removed from the YAML file.
+  ([#337](https://github.com/davep/blogmore/pull/337))
 
 ## v2.2.1